### 2025-10-07 85a763e
DOC: update JOSS2 paper.bib
