+++ docs/JOSS2/paper.bib
@@ -2,7 +2,8 @@
   title={Data-induced interactions of sparse sensors},
   author={Klishin, Andrei A and Kutz, J Nathan and Manohar, Krithika},
   journal={arXiv preprint arXiv:2307.11838},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2307.11838}
 }
 
 @article{karnik2024constrained,
@@ -10,14 +11,8 @@
   author={Karnik, Niharika and Abdo, Mohammad G and Estrada-Perez, Carlos E and Yoo, Jun Soo and Cogliati, Joshua J and Skifton, Richard S and Calderoni, Pattrick and Brunton, Steven L and Manohar, Krithika},
   journal={IEEE Sensors Journal},
   year={2024},
-  publisher={IEEE}
-}
-
-@article{klishin2025doubledescent,
-  title={Origins and mitigation of double descent in reduced order modeling},
-  author={Klishin, Andrei A and Otto, Samuel E and Kutz, J Nathan and Manohar, Krithika},
-  journal={In preparation},
-  year={2025}
+  publisher={IEEE},
+  doi={10.1109/JSEN.2024.3368875}
 }
 
 @article{erichson2020shallow,
@@ -28,7 +23,8 @@
   number={2238},
   pages={20200097},
   year={2020},
-  publisher={The Royal Society Publishing}
+  publisher={The Royal Society Publishing},
+  doi={10.1098/rspa.2020.0097}
 }
 
 @article{manohar2018predicting,
@@ -38,7 +34,8 @@
   volume={48},
   pages={87--95},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jmsy.2018.01.011}
 }
 
 @article{alonso2010novel,
@@ -49,7 +46,8 @@
   number={12},
   pages={4285--4295},
   year={2010},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TGRS.2010.2051231}
 }
 
 @article{krause2008near,
@@ -58,18 +56,8 @@
   journal={Journal of Machine Learning Research},
   volume={9},
   number={2},
-  year={2008}
-}
-
-@article{alexanderian2021optimal,
-  title={Optimal experimental design for infinite-dimensional Bayesian inverse problems governed by PDEs: A review},
-  author={Alexanderian, Alen},
-  journal={Inverse Problems},
-  volume={37},
-  number={4},
-  pages={043001},
-  year={2021},
-  publisher={IOP Publishing}
+  year={2008},
+  doi={10.5555/1390681.1390689}
 }
 
 @article{donoho2006compressed,
@@ -80,36 +68,21 @@
   number={4},
   pages={1289--1306},
   year={2006},
-  publisher={IEEE}
-}
-
-@article{summers2015submodularity,
-  title={On submodularity and controllability in complex dynamical networks},
-  author={Summers, Tyler H and Cortesi, Fabrizio L and Lygeros, John},
-  journal={IEEE Transactions on Control of Network Systems},
-  volume={3},
-  number={1},
-  pages={91--101},
-  year={2015},
-  publisher={IEEE}
-}
-
-@article{joshi2008sensor,
-  title={Sensor selection via convex optimization},
-  author={Joshi, Siddharth and Boyd, Stephen},
-  journal={IEEE Transactions on Signal Processing},
-  volume={57},
-  number={2},
-  pages={451--462},
-  year={2008},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TIT.2006.871582}
 }
 
 @article{de2021pysensors,
-  title={PySensors: A Python package for sparse sensor placement},
[…]
