### 2025-10-07 1ba35df
DOC: correct JOSS2 paper.bib DOI tags
