+++ docs/JOSS2/paper.bib
@@ -57,7 +57,7 @@
   volume={9},
   number={2},
   year={2008},
-  doi={10.5555/1390681.1390689}
+  url={https://dl.acm.org/doi/10.5555/1390681.1390689}
 }
 
 @article{donoho2006compressed,
@@ -139,16 +139,16 @@ doi={10.21105/joss.02828}
 }
 
 @article{yi2011optimal,
-author = {Yi, Ting-Hua and Li, Hong-Nan and Gu, Ming},
-title = {Optimal sensor placement for structural health monitoring based on multiple optimization strategies},
-journal = {The Structural Design of Tall and Special Buildings},
-volume = {20},
-number = {7},
-pages = {881-900},
-keywords = {structural health monitoring (SHM), optimal sensor placement (OSP), modal assurance criterion (MAC), matlab toolbox, Guangzhou New TV Tower (GNTVT)},
-doi = {https://doi.org/10.1002/tal.712},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/tal.712},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/tal.712},
-abstract = {SUMMARY Careful selection and placement of sensors are the critical issue in the construction and implementation of an effective structural health monitoring system. A hybrid method termed the optimal sensor placement strategy (OSPS) based on multiple optimization methods is proposed in this paper. The initial sensor placement is firstly obtained by the QR factorization. Then, using the minimization of the off-diagonal elements in the modal assurance criterion matrix as a measure of the utility of a sensor configuration, the quantity of the sensors is determined by the forward and backward sequential sensor placement algorithm together. Finally, the locations of the sensor are determined by the dual-structure coding-based generalized genetic algorithm (GGA). Taking the scientific calculation software matlab (MathWorks, Natick, MA, USA) as a platform, an OSPS toolbox, which is working as a black box, is developed based on the command-line compiling and graphical user interface-aided graphical interface design. The characteristic and operation method of the toolbox are introduced in detail, and the scheme selection of the OSP is carried out on the world's tallest TV tower (Guangzhou New TV Tower) based on the developed toolbox. The results indicate that the proposed method is effective and the software package has a friendly interface, plenty of functions, good expansibility and is easy to operate, which can be easily applied in practical engineering. Copyright © 2011 John Wiley \& Sons, Ltd.},
-year = {2011}
+author={Yi, Ting-Hua and Li, Hong-Nan and Gu, Ming},
+title={Optimal sensor placement for structural health monitoring based on multiple optimization strategies},
+journal={The Structural Design of Tall and Special Buildings},
+volume={20},
+number={7},
+pages={881-900},
+keywords={structural health monitoring (SHM), optimal sensor placement (OSP), modal assurance criterion (MAC), matlab toolbox, Guangzhou New TV Tower (GNTVT)},
+doi={10.1002/tal.712},
+url={https://onlinelibrary.wiley.com/doi/abs/10.1002/tal.712},
+eprint={https://onlinelibrary.wiley.com/doi/pdf/10.1002/tal.712},
[…]
