### 2025-10-08 #9161 @danielskatz
Thanks - I'll assign you, but in the future, you can also assign yourself with `@editorialbot assign me as editor`
