### 2025-10-13 33b105d
DOC: add funding acknowledgement in JOSS2 paper
