### 2025-10-13 84310b4
DOC: Update JOSS2 paper.md
