### 2025-10-15 #9161 @kanishkan91
@henrykironde ,  @kellyrowland  , @isdanni - Would any of you be interested in reviewing this paper for JOSS?
