### 2025-10-16 #9265 @editorialbot
[REVIEW]: PySensors 2.0: A Python Package for Sparse Sensor Placement
**Submitting author:** @niharika2999 (<a href="http://orcid.org/0000-0002-4259-0294">Niharika Karnik</a>)
**Repository:** https://github.com/dynamicslab/pysensors
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.2
