### 2025-10-22 1cf93c3
DOC: update references
