### 2026-01-11 884cdf0
DOC: update JOSS2 paper
