+++ docs/JOSS2/paper.md
@@ -54,25 +54,24 @@ bibliography: paper.bib
 # Statement of need
 The scalable optimization of sensor placement is critical for efficient monitoring, control, and decision-making in complex engineering systems. Sensor measurements are necessary for real-time estimation of high-dimensional fluid flows [@erichson2020shallow], large-scale flexible structures [@manohar2018predicting], temperature fields and pressure fields across geophysical [@alonso2010novel] and nuclear energy systems [@karnik2024constrained]. Accurate real-time tracking of key system variables depends sensitively on the locations of sensors deployed within the system, which have to be optimized for the desired task. In general, the selection of an optimal subset of sensors among candidate locations is NP-hard, necessitating heuristic or greedy approaches such as compressed sensing [@donoho2006compressed] or greedy algorithms that exploit submodularity [@krause2008near].
 
-![An overview image of capabilities of Pysensors](./Fig1.jpeg "PySensors 2.0 expands its capabilities by introducing custom basis functions, optimizers, constraints, solvers, and uncertainty quantification, enabling constrained sensing, over- and under-sampling,  and uncertainty quantification in the presence of noisy sensor measurements.")
+![An overview image of capabilities of PySensors](./Fig1.jpeg "PySensors 2.0 expands its capabilities by introducing custom basis functions, optimizers, constraints, solvers, and uncertainty quantification, enabling constrained sensing, over- and under-sampling,  and uncertainty quantification in the presence of noisy sensor measurements.")
 
-`PySensors` is a Python package [@de2021pysensors] dedicated to solving the complex challenge of optimal sensor placement in data-driven systems. It implements advanced sparse optimization algorithms that use dimensionality reduction techniques to identify the most informative measurement locations with remarkable efficiency [@manohar2018data;@brunton2016sparse;@clark2020multi]. It helps users identify the best locations for sensors when working with complex high dimensional data, focusing on both reconstruction [@manohar2018data] and classification [@brunton2016sparse] tasks. The package follows `scikit-learn` conventions for user-friendly access while offering advanced customization options for experienced users.  Other sensor placment packages such as `Chama` [@klise2017sensor], `Polire` [@narayanan2020toolkit], and `OSPS` toolbox [@yi2011optimal], focus primarily on event detection, Gaussian process modeling, and structural health monitoring respectively, while `PySensors` specifically targets signal reconstruction and classification applications.
+`PySensors` is a Python package [@de2021pysensors] dedicated to solving the complex challenge of optimal sensor placement in data-driven systems. It implements advanced sparse optimization algorithms that use dimensionality reduction techniques to identify the most informative measurement locations with remarkable efficiency [@manohar2018data;@brunton2016sparse;@clark2020multi]. It helps users identify the best locations for sensors when working with complex high dimensional data, focusing on both reconstruction [@manohar2018data] and classification [@brunton2016sparse] tasks. The package follows `scikit-learn` conventions for user-friendly access while offering advanced customization options for experienced users.  Other sensor placement packages such as `Chama` [@klise2017sensor], `Polire` [@narayanan2020toolkit], and `OSPS` toolbox [@yi2011optimal], focus primarily on event detection, Gaussian process modeling, and structural health monitoring respectively, while `PySensors` specifically targets signal reconstruction and classification applications.
 
[…]
