### 2026-01-21 8f31499
FIX: make CustomBasis sklearn compliant
