### 2026-01-28 4b6ac97
BLD: store all project metadata in pyproject.toml
