+++ requirements.txt
@@ -1,5 +1,0 @@
-scikit-learn>=0.23
-matplotlib
-numpy
-pandas
-scipy
