+++ requirements-examples.txt
@@ -1,6 +1,0 @@
-jupyter
-matplotlib
-netCDF4
-notebook
-pandas
-seaborn
