### 2026-01-28 c7c4f5b
BLD: move flake8 config from setup to pre-commit
