### 2026-01-28 82b4be4
BLD: update CI testing installation
