+++ .github/workflows/main.yml
@@ -63,7 +63,8 @@ jobs:
             python-version: ${{ matrix.python-version }}
         - name: Install dependencies
           run: |
-            pip install -r requirements-dev.txt
+            pip install --upgrade pip
+            pip install -e ".[dev,examples]"
         - name: Test with pytest
           run: |
             pytest tests --cov=pysensors --cov-report=xml
@@ -71,7 +72,7 @@ jobs:
           uses: codecov/codecov-action@v5
           with:
             token: ${{ secrets.CODECOV_TOKEN }}
-            file: ./coverage.xml
+            files: ./coverage.xml
         - name: Execute feature notebook with papermill
           run: |
             pip install papermill
@@ -80,6 +81,6 @@ jobs:
         - uses: actions/cache@v4
           with:
             path: ~/.cache/pip
-            key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements-dev.txt') }}
+            key: ${{ runner.os }}-pip-${{ hashFiles('**/pyproject.toml') }}
             restore-keys: |
               ${{ runner.os }}-pip-
