### 2026-01-28 3f257a8
DOC: update README installation guidelines
