### 2026-01-28 802c9ef
Rewrite pyproject.toml, remove setup.py and requirement files (#32)
* BLD: python3.9 is deprecated, require python>=3.10 upto 3.13

* BLD: store all project metadata in pyproject.toml

* BLD: move flake8 config from setup to pre-commit

* BLD: update CI testing installation

* DOC: update README installation guidelines
