+++ setup.cfg
@@ -1,20 +1,0 @@
-[flake8]
-exclude =
-    .git,
-    .venv,
-    dist,
-    build,
-    __pycache__
-ignore =
-    # Line break before binary operator - Conflicts black
-    W503
-    # Whitespace before ':' - Conflicts black
-    E203
-per-file-ignores =
-    __init__.py:F401,F403
-max-line-length = 88
-import-order-style = smarkets
-statistics = True
-count = True
-verbose = 1
-# format = [%(code)s] %(text)s @ %(path)s:%(row)d:%(col)d
