+++ .readthedocs.yml
@@ -7,9 +7,10 @@ build:
 
 python:
   install:
-    - requirements: requirements-dev.txt
     - method: pip
       path: .
+      extra_requirements:
+        - dev
 
 sphinx:
   configuration: docs/conf.py
