+++ .github/workflows/release.yml
@@ -14,11 +14,13 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: 3.11
-      - name: Install dependencies
-        run: pip install wheel
+          python-version: "3.12"
+      - name: Install Build
+        run: |
+          python -m pip install --upgrade pip
+          pip install build
       - name: Build package
-        run: python setup.py sdist bdist_wheel
+        run: python -m build .
       - name: Upload package
         uses: pypa/gh-action-pypi-publish@master
         with:
