+++ .github/workflows/main.yml
@@ -32,7 +32,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.10", "3.11", "3.12", 3.13]
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
 
     steps:
     - uses: actions/checkout@v4
@@ -53,7 +53,7 @@ jobs:
           fail-fast: false
           max-parallel: 4
           matrix:
-            python-version: ["3.10", "3.11", "3.12", 3.13]
+            python-version: ["3.10", "3.11", "3.12", "3.13"]
 
       steps:
         - uses: actions/checkout@v4
