### 2026-02-07 #9265 @niharika2999
Hello @kanishkan91 we have conducted a new Github release of the repository with the version number v0.4.2 (https://github.com/dynamicslab/pysensors/tree/v0.4.2).

However, as an outside collaborator to this repository, I don't have the necessary permissions to set up the Zenodo integration or manage the archival deposit. I'm currently coordinating with the repository owner to complete these steps.
We will ensure the Zenodo archive has:

Title matching the JOSS paper
Author list matching the paper order and names
ORCIDs where available
Correct license information

Sorry for the delay and I will keep you updated about the zenodo release. Thank you!
