### 2026-02-17 #9265 @danielskatz
👋 @niharika2999 - the zenodo link goes to a release of version v0.3.5, which doesn't match the version tag. And using what @kanishkan91 calls "the universal zenodo link" doesn't suffice for JOSS. We need a DOI that will resolve to the specific version.
