### 2026-02-17 a449ef7
minor changes in paper.md
