+++ docs/JOSS2/paper.md
@@ -32,15 +32,15 @@ authors:
     affiliation: 1
     orcid: 0000-0002-1582-6767
 affiliations:
- - name: Department of Mechanical Engineering, University of Washington
+ - name: Department of Mechanical Engineering, University of Washington, USA
    index: 1
- - name: Department of Applied Mathematics, University of Washington
+ - name: Department of Applied Mathematics, University of Washington, USA
    index: 2
- - name: Idaho National Laboratory
+ - name: Idaho National Laboratory, USA
    index: 3
- - name: Department of Mechanical Engineering, University of Hawai'i at Mānoa
+ - name: Department of Mechanical Engineering, University of Hawai'i at Mānoa, USA
    index: 4
- - name: Department of Biology, University of Washington
+ - name: Department of Biology, University of Washington, USA
    index: 5
 date: 12 August 2025
 bibliography: paper.bib
@@ -49,29 +49,29 @@ bibliography: paper.bib
 
 # Summary
 
-`PySensors` is a Python package for the optimization of sparse sensor placement for reconstruction and classification tasks. This major update to `PySensors` introduces novel spatial constraints for sensor placement, including the enforcement of maximum or exact sensor counts in specific regions, predetermined sensor locations, and minimum distances between sensors. This functionality is extended to support custom basis inputs, enabling integration of any data-driven or spectral basis.  We also propose a thermodynamic approach that goes beyond a single "optimal" sensor configuration and maps the complete landscape of sensor interactions induced by the training data. This comprehensive view facilitates integration with external selection criteria and enables assessment of sensor replacement impacts. The new optimization technique also accounts for over- and under-sampling of sensors, utilizing a regularized least squares approach for robust reconstruction. Additionally, we implement noise-induced uncertainty quantification of the estimation error and provide pointwise uncertainty heat maps to guide deployment decisions. To highlight these additions, we outline the mathematical algorithms and theory underlying these new capabilities. The usage of these new features is illustrated with code examples and practical advice for implementation across multiple application domains. Finally, we outline a roadmap of potential extensions to further strengthen the package's functionality and applicability to emerging sensing challenges.
+`PySensors` is a Python package for the optimization of sparse sensor placement for reconstruction and classification tasks. This major update to `PySensors` introduces novel spatial constraints for sensor placement, including the enforcement of maximum or exact sensor counts in specific regions, predetermined sensor locations, and minimum distances between sensors. This functionality is extended to support custom basis inputs, enabling integration of any data-driven or spectral basis.  We also use a thermodynamic approach that goes beyond a single "optimal" sensor configuration and maps the complete landscape of sensor interactions induced by the training data. This comprehensive view facilitates integration with external selection criteria and enables assessment of sensor replacement impacts. The new optimization technique also accounts for over- and under-sampling of sensors, utilizing a regularized least squares approach for robust reconstruction. Additionally, we implement noise-induced uncertainty quantification of the estimation error and provide pointwise uncertainty heat maps to guide deployment decisions. To highlight these additions, we outline the mathematical algorithms and theory underlying these new capabilities. The usage of these new features is illustrated with code examples and practical advice for implementation across multiple application domains. Finally, we outline a roadmap of potential extensions to further strengthen the package's functionality and applicability to emerging sensing challenges.
 
 # Statement of need
[…]
