### 2026-02-17 #9265 @danielskatz
@niharika2999 - Additionally, I've proofread the paper, and suggest the changes in https://github.com/dynamicslab/pysensors/pull/33 - please merge this, or let me know what you disagree with
