### 2026-02-17 8dc36e0
minor changes for JOSS publication (#33)
* minor changes in paper.md

* minor changes in paper.bib
