+++ docs/JOSS2/paper.bib
@@ -51,7 +51,7 @@
 }
 
 @article{krause2008near,
-  title={Near-optimal sensor placements in Gaussian processes: Theory, efficient algorithms and empirical studies.},
+  title={Near-optimal sensor placements in {G}aussian processes: Theory, efficient algorithms and empirical studies},
   author={Krause, Andreas and Singh, Ajit and Guestrin, Carlos},
   journal={Journal of Machine Learning Research},
   volume={9},
@@ -80,7 +80,7 @@ volume={6},
 number={58},
 pages={2828},
 author={de Silva, Brian M. and Manohar, Krithika and Clark, Emily and Brunton, Bingni W. and Kutz, J. Nathan and Brunton, Steven L.},
-title={PySensors: A Python package for sparse sensor placement},
+title={PySensors: A {P}ython package for sparse sensor placement},
 journal={Journal of Open Source Software},
 doi={10.21105/joss.02828}
 }
