### 2026-02-17 #9265 @niharika2999
Hello @kanishkan91 and @danielskatz,

Thank you for the minor changes and comments @danielskatz. We agree with all of them and have merged your PR with the main branch. After this, we have conducted a new release of the repository and also archived it in Zenodo. The new DOI is : https://doi.org/10.5281/zenodo.18665848.

Thank you so much!
