### 2026-02-17 #9265 @danielskatz
Congratulations to @niharika2999 (Niharika Karnik) and co-authors on your JOSS publication!!

JOSS depends completely on volunteers - we couldn't do this without you
